Album Cover Assigner
My ongoing project that seeks to replace my library of MP3s with ID3
tags that are broken or have been poorly assigned. My biggest problems
are the album covers, some of which are empty and some of which are not
assigned properly. This script will hopefully use a combination of the
song name, song artist and album name to search databases such as rdio
and last.fm to pull metadata about the song and assign them to their
proper catagories.